OPENEUROPA-3373: Refactoring tests.
