+++ src/Plugin/search_api/display/ListDisplayDeriver.php
@@ -27,7 +27,7 @@ class ListDisplayDeriver extends DisplayDeriverBase {
         $entity_type = $datasource->getEntityTypeId();
         $bundles = $datasource->getBundles();
         foreach ($bundles as $id => $label) {
-          $id = 'list_display' . PluginBase::DERIVATIVE_SEPARATOR . $entity_type . PluginBase::DERIVATIVE_SEPARATOR . $id;
+          $id = $entity_type . PluginBase::DERIVATIVE_SEPARATOR . $id;
           $definition = $base_plugin_definition;
           $definition['label'] = $this->t('List display %id', ['%id' => $id]);
           $definition['description'] = $this->t('List display %id', ['%id' => $id]);
